# src/docs/swagger.yaml
definitions:
...
  github_com_hvmidrezv_web-app_api_dto.CityResponse:
    properties:
      country:
        $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_hvmidrezv_web-app_api_dto.CountryResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CityResponse'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_hvmidrezv_web-app_api_dto.CreateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - countryId
    - name
    type: object
  github_com_hvmidrezv_web-app_api_dto.CreateUpdateCountryRequest:
    properties:
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  github_com_hvmidrezv_web-app_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  github_com_hvmidrezv_web-app_api_dto.PagedList-github_com_hvmidrezv_web-app_api_dto_CityResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CityResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_hvmidrezv_web-app_api_dto.PagedList-github_com_hvmidrezv_web-app_api_dto_CountryResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CountryResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_hvmidrezv_web-app_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.Sort'
        type: array
    type: object
...
  github_com_hvmidrezv_web-app_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
  github_com_hvmidrezv_web-app_api_dto.UpdateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    type: object
...
  github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_validations.ValidationError'
        type: array
    type: object
  github_com_hvmidrezv_web-app_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_hvmidrezv_web-app_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/cities/:
    post:
      consumes:
      - application/json
      description: Create a City
      parameters:
      - description: Create a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CreateCityRequest'
      responses:
        "201":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a City
      tags:
      - Cities
  /v1/cities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a City
      tags:
      - Cities
    get:
      consumes:
      - application/json
      description: Get a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a City
      tags:
      - Cities
    put:
      consumes:
      - application/json
      description: Update a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.UpdateCityRequest'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a City
      tags:
      - Cities
  /v1/cities/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Cities
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.PagedList-github_com_hvmidrezv_web-app_api_dto_CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Cities
      tags:
      - Cities
  /v1/countries/:
    post:
      consumes:
      - application/json
      description: Create a country
      parameters:
      - description: Create a country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CreateUpdateCountryRequest'
      responses:
        "201":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a country
      tags:
      - Countries
  /v1/countries/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a country
      tags:
      - Countries
    get:
      consumes:
      - application/json
      description: Get a country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a country
      tags:
      - Countries
    put:
      consumes:
      - application/json
      description: Update a country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CreateUpdateCountryRequest'
      responses:
        "200":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a country
      tags:
      - Countries
  /v1/countries/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Countries
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_hvmidrezv_web-app_api_dto.PagedList-github_com_hvmidrezv_web-app_api_dto_CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_hvmidrezv_web-app_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Countries
      tags:
      - Countries
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"